backend/requirements.txt

fastapi
uvicorn
supabase
google-genai
pydantic
